Add a new precheck method
In some situation the check block only makes sense if some preconditions
are met. One example are mountpoint options: If a certain mount (e.g.
`/tmp`) is no seperate mountpoint the check should fail, but we cannot
fix it (we cannot set mountoptions on a directory that is no mount).

These prechecks can now be definied with the precheck method.